Added bool for interpreter, added simple bool tests without branching or booleans ops yet

+++ compile.h
@@ -2,4 +2,4 @@
 
 #include "rep/r0.h"
 
-bool test_compile(const r0::P &p, int expect);
+bool test_compile(const r0::P &p, int64_t expect);
